unit tests from content_blocking.rs moved to tests/unit/content_blocking.rs
